# api/openapi/code-winch.yaml
openapi: 3.0.3
info:
  title: Code Winch API
  version: 1.0.0
  description: |
    Contract for managing runs and reading their durable event history. Request and
    response bodies never contain resolved credentials or sandbox execution handles.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: getHealth
      summary: Report service health
      security: []
      responses:
        "200":
          description: The service is healthy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /runs:
    post:
      operationId: createRun
      summary: Create a run
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: The run was created.
          headers:
            Location:
              description: Canonical URL of the created run.
              schema: { type: string }
            ETag:
              $ref: "#/components/headers/ETag"
            Idempotency-Key:
              $ref: "#/components/headers/IdempotencyKey"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "400": { $ref: "#/components/responses/ProblemBadRequest" }
        "401": { $ref: "#/components/responses/ProblemUnauthorized" }
        "409": { $ref: "#/components/responses/ProblemIdempotencyConflict" }
        "422": { $ref: "#/components/responses/ProblemValidationFailed" }
  /runs/{runId}:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: getRun
      summary: Read a run
      responses:
        "200":
          description: Current run representation.
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "401": { $ref: "#/components/responses/ProblemUnauthorized" }
        "404": { $ref: "#/components/responses/ProblemRunNotFound" }
  /runs/{runId}/start:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      operationId: startRun
      summary: Start a created run
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "202": { $ref: "#/components/responses/RunAccepted" }
        "400": { $ref: "#/components/responses/ProblemBadRequest" }
        "401": { $ref: "#/components/responses/ProblemUnauthorized" }
        "404": { $ref: "#/components/responses/ProblemRunNotFound" }
        "409": { $ref: "#/components/responses/ProblemStateConflict" }
        "412": { $ref: "#/components/responses/ProblemPreconditionFailed" }
        "422": { $ref: "#/components/responses/ProblemUnsupportedProfile" }
        "428": { $ref: "#/components/responses/ProblemPreconditionRequired" }
  /runs/{runId}/stop:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      operationId: stopRun
      summary: Stop a run
      description: Replaying this operation, including for a terminal run, is safe.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StopRunRequest" }
      responses:
        "202": { $ref: "#/components/responses/RunAccepted" }
        "400": { $ref: "#/components/responses/ProblemBadRequest" }
        "401": { $ref: "#/components/responses/ProblemUnauthorized" }
        "404": { $ref: "#/components/responses/ProblemRunNotFound" }
        "409": { $ref: "#/components/responses/ProblemIdempotencyConflict" }
        "412": { $ref: "#/components/responses/ProblemPreconditionFailed" }
        "428": { $ref: "#/components/responses/ProblemPreconditionRequired" }
  /runs/{runId}/events:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: listRunEvents
      summary: Read a page of durable run events
      parameters:
        - name: after_sequence
          in: query
          description: Return events with a sequence strictly greater than this value.
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Events ordered by ascending sequence.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventPage" }
        "400": { $ref: "#/components/responses/ProblemBadRequest" }
        "401": { $ref: "#/components/responses/ProblemUnauthorized" }
        "404": { $ref: "#/components/responses/ProblemRunNotFound" }
  /runs/{runId}/events/stream:
    parameters:
      - $ref: "#/components/parameters/RunId"
    get:
      operationId: streamRunEvents
      summary: Resume an ordered run event WebSocket stream
      description: |
        Upgrades to WebSocket after header or secure-cookie authentication and
        exact Origin validation. Messages are event, caught_up, heartbeat, or a
        resumable disconnect carrying lastSequence. Authorization is refreshed
        while the connection remains open.
      parameters:
        - name: after_sequence
          in: query
          description: Resume strictly after this durable sequence.
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
      responses:
        "101": { description: WebSocket protocol switch. }
        "400": { $ref: "#/components/responses/ProblemBadRequest" }
        "401": { $ref: "#/components/responses/ProblemUnauthorized" }
        "403": { description: The WebSocket Origin is not allowed. }
        "404": { $ref: "#/components/responses/ProblemRunNotFound" }
  /runs/{runId}/input:
    parameters:
      - $ref: "#/components/parameters/RunId"
    post:
      operationId: sendRunInput
      summary: Send idempotent input to a running run
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RunInputRequest" }
      responses:
        "202":
          description: Input was durably accepted or safely replayed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InputAccepted" }
        "400": { $ref: "#/components/responses/ProblemBadRequest" }
        "401": { $ref: "#/components/responses/ProblemUnauthorized" }
        "404": { $ref: "#/components/responses/ProblemRunNotFound" }
        "409": { $ref: "#/components/responses/ProblemStateConflict" }
        "412": { $ref: "#/components/responses/ProblemPreconditionFailed" }
        "428": { $ref: "#/components/responses/ProblemPreconditionRequired" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/RunId" }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key. Reuse with a different request returns an idempotency conflict.
      schema: { type: string, minLength: 1, maxLength: 128, pattern: "^[A-Za-z0-9._~-]+$" }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag from the most recently read run representation.
      schema: { type: string, pattern: '^"[^"\\r\\n]+"$' }
  headers:
    ETag:
      description: Strong validator representing the run's current version.
      schema: { type: string }
    IdempotencyKey:
      description: Idempotency key associated with the response.
      schema: { type: string }
  responses:
    RunAccepted:
      description: The command was accepted or safely replayed.
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
        Idempotency-Key: { $ref: "#/components/headers/IdempotencyKey" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Run" }
    ProblemBadRequest:
      description: The request could not be parsed.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/bad-request", title: "Bad request", status: 400, code: "bad_request", detail: "The request body is not valid JSON.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemUnauthorized:
      description: Authentication is required or invalid.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/unauthorized", title: "Unauthorized", status: 401, code: "unauthorized", detail: "Valid authentication is required.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemRunNotFound:
      description: The run does not exist or is not visible to the caller.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/run-not-found", title: "Run not found", status: 404, code: "run_not_found", detail: "The requested run was not found.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemStateConflict:
      description: The requested transition is invalid for the current run state.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/run-state-conflict", title: "Run state conflict", status: 409, code: "run_state_conflict", detail: "The run cannot be started from its current state.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemIdempotencyConflict:
      description: The idempotency key was previously used for a different request.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/idempotency-conflict", title: "Idempotency conflict", status: 409, code: "idempotency_conflict", detail: "Use a new Idempotency-Key for a different request.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemPreconditionFailed:
      description: The supplied ETag no longer represents the run.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/precondition-failed", title: "Precondition failed", status: 412, code: "precondition_failed", detail: "Read the run and retry with its current ETag.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemPreconditionRequired:
      description: If-Match is required for this command.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/precondition-required", title: "Precondition required", status: 428, code: "precondition_required", detail: "Supply the run's current ETag in If-Match.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemUnsupportedProfile:
      description: The run's harness and sandbox profile pair is not supported by this deployment.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/unsupported-profile", title: "Unsupported profile", status: 422, code: "unsupported_profile", detail: "This deployment cannot run the run's harness and sandbox profile pair.", requestId: "01K0EXAMPLE00000000000000" } } }
    ProblemValidationFailed:
      description: One or more request fields are invalid.
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" }, example: { type: "https://code-winch.dev/problems/validation-failed", title: "Validation failed", status: 422, code: "validation_failed", detail: "Correct the invalid fields and retry.", requestId: "01K0EXAMPLE00000000000000", errors: [{ field: "workspacePath", code: "required", message: "workspacePath is required" }] } } }
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties: { status: { type: string, enum: [ok] } }
    RunId: { type: string, pattern: "^[0-9A-HJKMNP-TV-Z]{26}$", example: "01J00000000000000000000000" }
    RunState: { type: string, enum: [created, queued, preparing, running, stopping, completed, failed, cancelled] }
    CreateRunRequest:
      type: object
      additionalProperties: false
      required: [workspacePath, harnessProfile, sandboxProfile]
      properties:
        workspacePath: { type: string, minLength: 1, maxLength: 4096, description: Workspace reference; it must not contain credentials. }
        harnessProfile: { type: string, minLength: 1, maxLength: 128 }
        sandboxProfile: { type: string, minLength: 1, maxLength: 128 }
    StopRunRequest:
      type: object
      additionalProperties: false
      properties:
        reason: { type: string, maxLength: 500 }
    RunInputRequest:
      type: object
      additionalProperties: false
      required: [kind]
      properties:
        kind: { type: string, enum: [text, interrupt, terminal_bytes, resize] }
        text: { type: string, maxLength: 65536 }
        bytes: { type: string, format: byte, maxLength: 87384 }
        rows: { type: integer, minimum: 1, maximum: 65535 }
        columns: { type: integer, minimum: 1, maximum: 65535 }
    InputAccepted:
      type: object
      additionalProperties: false
      required: [commandId, runId, kind, accepted]
      properties:
        commandId: { type: string }
        runId: { $ref: "#/components/schemas/RunId" }
        kind: { type: string, enum: [text, interrupt, terminal_bytes, resize] }
        accepted: { type: boolean }
    Run:
      type: object
      additionalProperties: false
      required: [id, state, version, createdAt, updatedAt, workspacePath, harnessProfile, sandboxProfile]
      properties:
        id: { $ref: "#/components/schemas/RunId" }
        state: { $ref: "#/components/schemas/RunState" }
        version: { type: integer, format: int64, minimum: 1, description: Monotonic concurrency version represented by ETag. }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        workspacePath: { type: string }
        harnessProfile: { type: string }
        sandboxProfile: { type: string }
        lastSequence: { type: integer, format: int64, minimum: 0, default: 0 }
    Event:
      type: object
      additionalProperties: false
      required: [eventId, runId, sequence, occurredAt, kind, schemaVersion, source, sensitivity, payload]
      properties:
        eventId: { type: string }
        runId: { $ref: "#/components/schemas/RunId" }
        sequence: { type: integer, format: int64, minimum: 1 }
        occurredAt: { type: string, format: date-time }
        kind: { type: string, minLength: 1 }
        schemaVersion: { type: integer, minimum: 1 }
        source: { type: object, additionalProperties: true }
        sensitivity: { type: string, enum: [public, operational, user-content, confidential, secret] }
        payload: { type: object, additionalProperties: true }
        extensions: { type: object, additionalProperties: true }
    EventPage:
      type: object
      additionalProperties: false
      required: [events, nextAfterSequence, hasMore]
      properties:
        events: { type: array, maxItems: 200, items: { $ref: "#/components/schemas/Event" } }
        nextAfterSequence: { type: integer, format: int64, minimum: 0 }
        hasMore: { type: boolean }
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, detail, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, description: Stable machine-readable error code. }
        detail: { type: string, description: Actionable summary that contains no content or secrets. }
        requestId: { type: string, description: Correlation identifier safe to include in logs. }
        errors: { type: array, items: { $ref: "#/components/schemas/FieldError" } }